requirements.txt
requests
openai
pandas
python-dotenv
